requirements.txt
# requirements.txt
# Python 3.11+

# GUI
PySide6>=6.6.0

# Async networking
aiohttp>=3.9.0
asyncssh>=2.14.0

# Modbus
pymodbus>=3.6.0

# Configuration
pyyaml>=6.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-qt>=4.2.0

# Utils
python-dateutil>=2.8.2